Backend/requirements.txt
fastapi
uvicorn[standard]
crewai==0.28.2
crewai-tools==0.0.12
pydantic-ai[logfire]==0.1.2
redis
httpx
langchain-community==0.0.31 